Add draft of lexer example #9
This adds an initial draft of the lexer example.  The underlying
transformation is not yet truly incrementaly, but its enough to
demonstrate the main concepts.

Have added notion of `Sequence` which extends `Index` by including a
`len()` function.  Unclear whether or not this is useful in the end.

This uses the spelling of tokenise which follows the british convention
because I'm british.  This also renames `Lexer` to `Tokenisation` which
I believe provides better terminology.  We have a tokenisation which is
updated incrementally, etc.