Implement PartiallyTransformable for Tokenisation
This puts in place the potential for creating a truly incremental
tokenisation.  At this stage, upon applying a transform, it simply
recomputes the meta-data from scratch how.  Still, this is now very
close.  There are a large number of questions arising from this,
however:

(1) Error State.  Currently, a transformation can fail with an error
(i.e. because the delta was invalid).  An interesting question is what
this means for the internal state of the Tokenisation.

(2) Error Recovery.  There is no mechanism for error recovery, though
this potentially could be supported.

(3) When a transformation fails, there is no current requirement to
check the return value.  I think adding that might be important.