Add Delta::iter() #19
This adds the ability to iterate over the rewrites contained within a
Delta.  Furthermore, we can map a rewrite from one type to another as
well.  So, for example, we can take a delta over integers and construct
a "mirror" delta over booleans.  This is quite useful for managing the
updates of meta-data.  However, its not clear that its actually useful
long term.