Implement `VecDelta`
The core data structure of this library appears to be `VecDelta`.  This
encodes zero or more rewrites in a given package, such that we can apply
them to an existing sequence.

Using the `longest_common_subsequence()` algorithm, we can obtain the diff
of two slices (represented as a `VecDelta`).  Then, using this, we can
convert the former into the latter.  However, from the perspective of
this library's intended uses, it is the information encoded within
`VecDelta` that is most interesting.  This is the package that will be
passed around to various incremental algorithms.

This includes various tests for `VecDelta` and the
`longest_common_subsequence()` algorithm.  The next step is really to
consider the _uses_ of this data structure.  Specifically, to think
about what a simple incremental parser might look like.